move resources logic from comms
