--- cli/medperf/comms/cube_assets/__init__.py
@@ -1,3 +1,0 @@
-from .download import download_resource
-
-__all__ = [download_resource]
