+++ cli/medperf/comms/entity_resources/utils.py
